feat: implement admin role management and permissions

+++ server/src/user_management_context/application/use_cases/delete_user_use_case.py
@@ -1,6 +1,6 @@
 from user_management_context.application.interfaces import UserRepository
 from user_management_context.application.commands import DeleteUserCommand
-from shared_kernel.authentication import NotAdminError, AuthenticatedUser
+from shared_kernel.authentication import AdminPermissionChecker
 
 
 class DeleteUserUseCase:
@@ -8,11 +8,7 @@ class DeleteUserUseCase:
         self.user_repository = user_repository
 
     def execute(self, command: DeleteUserCommand) -> None:
-        self._check_admin_permission(command.requesting_user)
+        AdminPermissionChecker.ensure_admin(command.requesting_user, "delete users")
         user_id = command.user_id
 
         self.user_repository.delete(user_id)
-
-    def _check_admin_permission(self, requesting_user: AuthenticatedUser) -> None:
-        if "admin" not in requesting_user.roles:
-            raise NotAdminError("Only administrators can delete users")
